requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
ciso8601==2.3.0
click==8.1.3
cloudpickle==2.2.1
cryptography==39.0.0
cx-Oracle==8.3.0
dask==2023.1.1
db-dtypes==1.0.5
decorator==5.1.1
frozenlist==1.3.3
fsspec==2023.1.0
gcsfs==2023.1.0
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.8.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-bigquery==3.4.2
google-cloud-bigquery-storage==2.18.1
google-cloud-core==2.3.2
google-cloud-error-reporting==1.8.0
google-cloud-logging==3.4.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.51.1
grpcio-status==1.51.1
idna==3.4
locket==1.0.0
multidict==6.0.4
numpy==1.24.1
oauthlib==3.2.2
oracledb==1.2.2
packaging==23.0
pandas==1.5.3
pandas-gbq==0.19.1
partd==1.3.0
proto-plus==1.22.2
protobuf==4.21.12
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydata-google-auth==1.6.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.0
toolz==0.12.0
typing-extensions==4.4.0
urllib3==1.26.14
yarl==1.8.2
google-cloud-secret-manager==2.10.0